Code clean up Added CI: Gitlab actions config
