+++ .github/workflows/ci.yml
@@ -1,0 +1,32 @@
+name: Continuous Integration
+
+on: [pull_request]
+
+jobs:
+  quality:
+    name: Quality checks
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout
+
+      - name: Composer normalize
+        run: composer normalize
+
+      - name: Install dependencies
+        run: composer install --prefer-dist --no-progress --no-suggest
+
+      - name: PHP Code Sniffer
+        if: ${{ always() }}
+        run: vendor/bin/phpcs
+
+      - name: PHP Mess Detector
+        if: ${{ always() }}
+        run: vendor/bin/phpmd src/ text cleancode,codesize,controversial
+
+      - name: PHP Stan
+        if: ${{ always() }}
+        run: vendor/bin/phpstan analyse -c phpstan.neon
+
+      - name: PHP Unit tests
+        if: ${{ always() }}
+        run: bin/phpunit
