--- .styleci.yml
@@ -1,6 +1,0 @@
-preset: laravel
-risky: true
-finder:
-  exclude:
-    - vendor
-  name: "*.php"
