+++ phpcs.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0"?>
+
+<ruleset  name="PHP_CodeSniffer">
+
+    <description>PHPCS configuration file.</description>
+    <file>src</file>
+    <file>tests</file>
+
+    <rule  ref="PSR12"/>
+</ruleset>
