+++ phpstan.neon
@@ -1,0 +1,5 @@
+parameters:
+	level: 6
+	paths:
+		- src
+		- tests
