+++ tests/unit/StatementClausesTest.php
@@ -27,7 +27,7 @@ class StatementClausesTest extends TestCase
             "C"
           ]
         }')->get();
-        self::assertEquals('LET x = @1_1', $result->query);
+        self::assertEquals('LET x = @'.$result->getQueryId().'_1', $result->query);
 
         $qb = new QueryBuilder();
         $object = new \stdClass();
@@ -39,13 +39,13 @@ class StatementClausesTest extends TestCase
         $result = $qb->let('x', $object)
             ->get();
         self::assertEquals(
-            'LET x = {"name":"Catelyn","surname":"Stark","alive":false,"age":40,"traits":[@1_1,@1_2,@1_3]}',
+            'LET x = {"name":"Catelyn","surname":"Stark","alive":false,"age":40,"traits":[@'.$result->getQueryId().'_1,@'.$result->getQueryId().'_2,@'.$result->getQueryId().'_3]}',
             $result->query
         );
 
         $result = (new QueryBuilder())->let('x', 'y')
             ->get();
-        self::assertEquals('LET x = @1_1', $result->query);
+        self::assertEquals('LET x = @'.$result->getQueryId().'_1', $result->query);
     }
 
     /**
@@ -65,7 +65,7 @@ class StatementClausesTest extends TestCase
             "C"
           ]
         }', 'Characters')->get();
-        self::assertEquals('INSERT @1_1 IN Characters', $result->query);
+        self::assertEquals('INSERT @'.$result->getQueryId().'_1 IN Characters', $result->query);
     }
 
     /**
@@ -81,7 +81,7 @@ class StatementClausesTest extends TestCase
                 '{ name: CONCAT(u.firstName, " ", u.lastName) }',
                 'users'
             )->get();
-        self::assertEquals('FOR u IN users UPDATE u WITH @1_1 IN users', $result->query);
+        self::assertEquals('FOR u IN users UPDATE u WITH @'.$result->getQueryId().'_1 IN users', $result->query);
     }
 
     /**
@@ -98,7 +98,7 @@ class StatementClausesTest extends TestCase
                 'users'
             )
             ->get();
-        self::assertEquals('FOR u IN users REPLACE u WITH @1_1 IN users', $result->query);
+        self::assertEquals('FOR u IN users REPLACE u WITH @'.$result->getQueryId().'_1 IN users', $result->query);
     }
 
     /**
@@ -122,7 +122,7 @@ class StatementClausesTest extends TestCase
             ->for('i', '1..1000')
             ->remove('{ _key: CONCAT(\'test\', i) }', 'users')
             ->get();
-        self::assertEquals('FOR i IN 1..1000 REMOVE @1_1 IN users', $result->query);
+        self::assertEquals('FOR i IN 1..1000 REMOVE @'.$result->getQueryId().'_1 IN users', $result->query);
     }
 
     /**
@@ -138,7 +138,7 @@ class StatementClausesTest extends TestCase
                 '{ logins: OLD.logins + 1 }',
                 'users'
             )->get();
-        self::assertEquals('UPSERT @1_1 INSERT @1_2 UPDATE @1_3 IN users', $result->query);
+        self::assertEquals('UPSERT @'.$result->getQueryId().'_1 INSERT @'.$result->getQueryId().'_2 UPDATE @'.$result->getQueryId().'_3 IN users', $result->query);
 
         $result = (new QueryBuilder())
             ->upsert(
@@ -148,7 +148,7 @@ class StatementClausesTest extends TestCase
                 'users',
                 true
             )->get();
-        self::assertEquals('UPSERT @1_1 INSERT @1_2 REPLACE @1_3 IN users', $result->query);
+        self::assertEquals('UPSERT @'.$result->getQueryId().'_1 INSERT @'.$result->getQueryId().'_2 REPLACE @'.$result->getQueryId().'_3 IN users', $result->query);
     }
 
     public function testUpdateMaintainsNullValue()
